Major revision of amendments
Reduce the number of changes that do not address issues of compliance
with SEA policies.  Some changes that do not address such issues have
been retained or added to correct grammatical or spelling errors, to
address issues of compliance with law, to improve clarity, to simplify
(by relying more on the parliamentary authority where possible), etc.

Add or correct cross-reference hyperlinks as appropriate.

Fully type out all amendments in accordance with requirements of the
Constitution.